mito-ai: refactor provider tests
